Connections between users fixed
I added new table (Connection) which holds the ids of the two users,
because the connection is bidirectional. Every user has many
connections, and a connection has two users.

I also added ConnectionRequestsController with services:
1) Send connection request
2) Accept connection request
3) Decline connection request
4) GetAll connection request for the current users

Everything seems to work fine. I will test it and finally be able to
focus only on the Android client :D